Add files via upload
The Customer class
In the context of our problem domain, what must the Customer class know? The description
directly mentions the following items, which are all attributes of a customer:
• The customer’s name
• The customer’s address
• The customer’s telephone number
These are all values that can be represented as strings and stored in the class’s fields. The
Customer class can potentially know many other things. One mistake that can be made at
this point is to identify too many things that an object is responsible for knowing. In some
applications, a Customer class might know the customer’s email address. This particular
problem domain does not mention that the customer’s email address is used for any purpose,
so we should not include it as a responsibility.
Now let’s identify the class’s methods. In the context of our problem domain, what must the
Customer class do? The only obvious actions are as follows:
• Create an object of the Customer class
• Set and get the customer’s name
• Set and get the customer’s address
• Set and get the customer’s telephone number
From this list we can see that the Customer class will have a constructor, as well as accessor
and mutator methods for each of its fields. Figure 6-29 shows a UML diagram for the
Customer class.
Customer
– name : String
– address : String
– phone : String
+ Customer()
+ setName(n : String) : void
+ setAddress(a : String) : void
+ setPhone(p : String) : void
+ getName() : String
+ getAddress() : String
+ getPhone() : String